update 1. Add lighthouse report in readme 2. Add browser change instruction gif
