+++ README.md
@@ -5,6 +5,8 @@ https://codemayank.github.io/stocks-app
 
 Note: For this site to work correctly in your browser please enable allow insecure content for this site in your browser. This is because the app uses websockets to fetch the latest stock data and the websocket server is running on an inscure endpoint.
 
+![](./screenshots/Browser_setting.gif)
+
 Screenshots
 Dark Mode
 ![](./screenshots/darkMode_ss.png)
@@ -13,7 +15,7 @@ Light Mode
 
 
 To run this app locally clone the repo and run
-`yarn start`
+`yarn install && yarn start`
 
 ## This project was built using
 * React
@@ -21,6 +23,8 @@ To run this app locally clone the repo and run
 * Create React App
 * Chakra Ui
 
-
+## Lighthouse Report
+![](./Lighthouse/lighthouse_report.png)
+[Full Report](./Lighthouse/lighthouse_report.json)
